feat: 후기 작성시 대상의 후기 고도 점수 갱신

+++ src/main/java/com/ll/olol/boundedContext/member/entity/Member.java
@@ -86,7 +86,7 @@ public class Member {
     @OneToMany(mappedBy = "member")
     private List<RecruitmentPeople> recruitmentPeople;
 
-    private double reviewScore;
+    private int reviewScore;
 
 
 //    @OneToMany(mappedBy = "fromMember", cascade = {CascadeType.ALL})
